Backend/requirements.txt
flask
flask-cors
opencv-python
numpy
pytesseract
Pillow
scikit-learn
werkzeug